Adds the chronos package
This package holds a Clock type that can be controlled to return
specific time values, which is useful for testing. It also provides a
`Measure` method that times the performance of a given operation, which
is a common clock task in our buildpacks.

This Clock type can be used as a drop-in replacement for the Clocks that
have been duplicated across our existing codebases.