Enables the tar decompression the ability to decompress a file that has no directory headers
- Previously vacation was given a tar file that had no directory
headers, which can be generated by handing the tar command a list of all
the files that should be tared, vacation would not be able to handle it
because it would try and create files in directories that did not exist.

- We now log all of the directories that have been created previously
and check to see that any file we create is within one of those
directories, if it is not then we create the directory for the file
before we try and write the file.