Promote BuildResult fields to LaunchMetadata
BuildResult fields Processes, Slices, and Labels will now be located on
a subfield called Launch. This separates them more distinctly from the
other fields on the BuildResult (Layers and Plan). Furthermore, it will
ease addition of support for Bill-of-Materials changes coming in
Buildpack Spec v0.5.

This deprecates the existing Processes, Slices, and Labels.