Adds EnvironmentVariables function to scribe.Emitter
- This function prints out a formatted map of the build and launch
environment variables and merges the shared environment variables into
both the build and launch environment for printing
- Refactors emitter unit tests to both use occam ContainLines matcher as
well as restructured to follow a pattern of delclaring top level
environment variable and doing neccessary configuration in it.Before(s)
instead of in the it() blocks themselves. This is an attempt to make the
conditions that the test is testing clearer as you have a well defined
set a parameters for all tests and then the modification is laid out
just before testing.