Add support for supplying strip-components via buildpack.toml
This PR adds a strip-components flag to postal’s service resolver. So that when it parses the buildpack tomls it can use artifacts that are tar’d in a slightly different manner. This would allow for customization of paketo forks that only have a different buildpack.toml.

For more details see https://paketobuildpacks.slack.com/archives/CULAS8ACD/p1619441959093300

Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>