Makes api changes to Executable and NopArchive
- NopArchive will now no longer write directly to the destination if a
name is not given becuase it now set a default name (`artifact`). If you
were using the generic Archive type you will see no difference but those
who were using the type directly must now specify a directory as the
destination, not a file name.
- NewExecutable now conforms to the same api as the other archive
instantiation functions excepting only a reader and allowing individuals
to set the file name using a `WithName()` option with mirrors the
behavior of all other archive objects.
-Executable now has a default file name (`artifact`) if you were using
the generic Archive type you will see no difference however becuase the
Executable instantiation function no longer requires a name a default is
now assigned on instantiation the same way it is done for the
NopArchive.
- Excutable no longer places the executable in a directory name `bin`
nor does it create all directories in the destination path. I felt that
automatically placing the executable in a `bin` directory felt to
opinionated, I think that it is more than reasonable for a user to want
to download and executable and not want to have it placed in a directory
named `bin` and this default behavior hinders that user. I think that it
is more than reasonable to expect that the destination given is a `bin`
directory if that is the desired effect. As for no longer creating all
of directories in the destination path, NopArchive is precedent setting
as it expects to write files only to destinations that exist already.