Adds the biome package
The initial implementation of this package includes functionality to dry
up a common pattern that I see in all buildpacks in Paketo that are
compatible with `watchexe`. The same function has propogated it's way
into almost every buildpack using `packit`. An example can be seen here
https://github.com/paketo-buildpacks/go-build/blob/e210480ead2699ea64b273578100b667f094e50c/detect.go#L57-L66

Given the switch to using environment variables as configuration is
almost complete across all buildpacks inside of the Paketo project it
seems like having a shared library for the parsing of environment
variable may come in handy.